% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{euclidean_distance_interface}
\alias{euclidean_distance_interface}
\alias{euclidean_distance}
\title{Calculate Euclidean Distance}
\usage{
euclidean_distance(point1, point2, p)
}
\arguments{
\item{point1}{A numeric vector representing the coordinates of the first point.}

\item{point2}{A numeric vector representing the coordinates of the second point.}

\item{p}{An integer specifying the number of dimensions or attributes in the data.}
}
\value{
A numeric value representing the Euclidean distance between \code{point1} and \code{point2}.
}
\description{
This function calculates the Euclidean distance between two points in a given space.
}
\examples{
# Example 1: Calculate Euclidean distance between two points in 2D space
point1 <- c(1.2, 2.3)
point2 <- c(4.5, 5.6)
p <- length(point1)
distance <- euclidean_distance(point1, point2, p)
cat("Euclidean Distance:", distance, "\n")

# Example 2: Calculate Euclidean distance between two points in 3D space
point1 <- c(1.2, 2.3, 3.1)
point2 <- c(4.5, 5.6, 6.7)
p <- length(point1)
distance <- euclidean_distance(point1, point2, p)
cat("Euclidean Distance:", distance, "\n")
}
