% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmm.R
\name{fitGMM_interface}
\alias{fitGMM_interface}
\alias{fitGMM}
\title{Fit a Gaussian Mixture Model (GMM)}
\usage{
fitGMM(data, K, maxIter)
}
\arguments{
\item{data}{A numeric vector containing the data points to fit the GMM to.}

\item{K}{An integer specifying the number of components in the GMM.}

\item{maxIter}{An integer specifying the maximum number of iterations for the EM algorithm.}
}
\value{
A matrix of responsibilities indicating the probability of each data point belonging to each component.
}
\description{
This function fits a Gaussian Mixture Model to the input data using C++ implementation.
}
\examples{
data <- matrix(rnorm(900), ncol = 3)
K <- 4
maxIter <- 10
result <- fitGMM(data, K, maxIter)
print(result)
}
